+++ src/android/FileTransfer.java
@@ -53,7 +53,6 @@ import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CordovaResourceApi;
 import org.apache.cordova.CordovaResourceApi.OpenForReadResult;
-import org.apache.cordova.LOG;
 import org.apache.cordova.PluginManager;
 import org.apache.cordova.PluginResult;
 import org.apache.cordova.Whitelist;
@@ -64,6 +63,7 @@ import org.json.JSONObject;
 
 import android.net.Uri;
 import android.os.Build;
+import android.util.Log;
 import android.webkit.CookieManager;
 
 public class FileTransfer extends CordovaPlugin {
@@ -277,7 +277,7 @@ public class FileTransfer extends CordovaPlugin {
      * @return FileUploadResult containing result of upload request
      */
     private void upload(final String source, final String target, JSONArray args, CallbackContext callbackContext) throws JSONException {
-        LOG.d(LOG_TAG, "upload " + source + " to " +  target);
+        Log.d(LOG_TAG, "upload " + source + " to " +  target);
 
         // Setup the options
         final String fileKey = getArgument(args, 2, "file");
@@ -291,26 +291,32 @@ public class FileTransfer extends CordovaPlugin {
         final JSONObject headers = args.optJSONObject(8) == null ? params.optJSONObject("headers") : args.optJSONObject(8);
         final String objectId = args.getString(9);
         final String httpMethod = getArgument(args, 10, "POST");
+		final int fileSize = args.getInt(11);
 
         final CordovaResourceApi resourceApi = webView.getResourceApi();
 
-        LOG.d(LOG_TAG, "fileKey: " + fileKey);
-        LOG.d(LOG_TAG, "fileName: " + fileName);
-        LOG.d(LOG_TAG, "mimeType: " + mimeType);
-        LOG.d(LOG_TAG, "params: " + params);
-        LOG.d(LOG_TAG, "trustEveryone: " + trustEveryone);
-        LOG.d(LOG_TAG, "chunkedMode: " + chunkedMode);
-        LOG.d(LOG_TAG, "headers: " + headers);
-        LOG.d(LOG_TAG, "objectId: " + objectId);
-        LOG.d(LOG_TAG, "httpMethod: " + httpMethod);
+        Log.d(LOG_TAG, "fileKey: " + fileKey);
+        Log.d(LOG_TAG, "fileName: " + fileName);
+        Log.d(LOG_TAG, "mimeType: " + mimeType);
+        Log.d(LOG_TAG, "params: " + params);
+        Log.d(LOG_TAG, "trustEveryone: " + trustEveryone);
+        Log.d(LOG_TAG, "chunkedMode: " + chunkedMode);
+        Log.d(LOG_TAG, "headers: " + headers);
+        Log.d(LOG_TAG, "objectId: " + objectId);
+        Log.d(LOG_TAG, "httpMethod: " + httpMethod);
+        Log.d(LOG_TAG, "fileSize: " + fileSize);
 
         final Uri targetUri = resourceApi.remapUri(Uri.parse(target));
+        // Accept a path or a URI for the source.
+        Uri tmpSrc = Uri.parse(source);
+        final Uri sourceUri = resourceApi.remapUri(
+            tmpSrc.getScheme() != null ? tmpSrc : Uri.fromFile(new File(source)));
 
         int uriType = CordovaResourceApi.getUriType(targetUri);
         final boolean useHttps = uriType == CordovaResourceApi.URI_TYPE_HTTPS;
         if (uriType != CordovaResourceApi.URI_TYPE_HTTP && !useHttps) {
             JSONObject error = createFileTransferError(INVALID_URL_ERR, source, target, null, 0, null);
-            LOG.e(LOG_TAG, "Unsupported URI: " + targetUri);
+            Log.e(LOG_TAG, "Unsupported URI: " + targetUri);
             callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
             return;
         }
@@ -325,19 +331,11 @@ public class FileTransfer extends CordovaPlugin {
                 if (context.aborted) {
                     return;
                 }
-
-                // We should call remapUri on background thread otherwise it throws
-                // IllegalStateException when trying to remap 'cdvfile://localhost/content/...' URIs
-                // via ContentFilesystem (see https://issues.apache.org/jira/browse/CB-9022)
-                Uri tmpSrc = Uri.parse(source);
-                final Uri sourceUri = resourceApi.remapUri(
-                        tmpSrc.getScheme() != null ? tmpSrc : Uri.fromFile(new File(source)));
-
                 HttpURLConnection conn = null;
                 HostnameVerifier oldHostnameVerifier = null;
                 SSLSocketFactory oldSocketFactory = null;
                 int totalBytes = 0;
-                int fixedLength = -1;
+				int fixedLength = fileSize;
                 try {
                     // Create return object
                     FileUploadResult result = new FileUploadResult();
@@ -404,7 +402,7 @@ public class FileTransfer extends CordovaPlugin {
                             }
                         }
                     } catch (JSONException e) {
-                        LOG.e(LOG_TAG, e.getMessage(), e);
+                        Log.e(LOG_TAG, e.getMessage(), e);
                     }
 
                     beforeData.append(LINE_START).append(BOUNDARY).append(LINE_END);
@@ -426,24 +424,26 @@ public class FileTransfer extends CordovaPlugin {
                         progress.setLengthComputable(true);
                         progress.setTotal(fixedLength);
                     }
-                    LOG.d(LOG_TAG, "Content Length: " + fixedLength);
+                    else {
+                        Log.d(LOG_TAG, "readResult length negative, using provided length");
+                        if (multipartFormUpload)
+                            fixedLength += stringLength;
+                        progress.setLengthComputable(true);
+                        progress.setTotal(fixedLength);
+                    }
+                    Log.d(LOG_TAG, "Content Length: " + fixedLength);
+					// Commented out, we don't care to support this
                     // setFixedLengthStreamingMode causes and OutOfMemoryException on pre-Froyo devices.
                     // http://code.google.com/p/android/issues/detail?id=3164
                     // It also causes OOM if HTTPS is used, even on newer devices.
-                    boolean useChunkedMode = chunkedMode || (Build.VERSION.SDK_INT < Build.VERSION_CODES.FROYO);
-                    useChunkedMode = useChunkedMode || (fixedLength == -1);
 
-                    if (useChunkedMode) {
+                    if (chunkedMode) {
                         conn.setChunkedStreamingMode(MAX_BUFFER_SIZE);
                         // Although setChunkedStreamingMode sets this header, setting it explicitly here works
                         // around an OutOfMemoryException when using https.
                         conn.setRequestProperty("Transfer-Encoding", "chunked");
                     } else {
                         conn.setFixedLengthStreamingMode(fixedLength);
-
-                        if (useHttps) {
-                            LOG.w(LOG_TAG, "setFixedLengthStreamingMode could cause OutOfMemoryException - switch to chunkedMode=true to avoid it if this is an issue.");
-                        }
                     }
 
                     conn.connect();
@@ -479,7 +479,7 @@ public class FileTransfer extends CordovaPlugin {
                             sendStream.write(buffer, 0, bytesRead);
                             if (totalBytes > prevBytesRead + 102400) {
                                 prevBytesRead = totalBytes;
-                                LOG.d(LOG_TAG, "Uploaded " + totalBytes + " of " + fixedLength + " bytes");
+                                Log.d(LOG_TAG, "Uploaded " + totalBytes + " of " + fixedLength + " bytes");
                             }
                             bytesAvailable = readResult.inputStream.available();
                             bufferSize = Math.min(bytesAvailable, MAX_BUFFER_SIZE);
@@ -505,13 +505,13 @@ public class FileTransfer extends CordovaPlugin {
                     synchronized (context) {
                         context.connection = null;
                     }
-                    LOG.d(LOG_TAG, "Sent " + totalBytes + " of " + fixedLength);
+                    Log.d(LOG_TAG, "Sent " + totalBytes + " of " + fixedLength);
 
                     //------------------ read the SERVER RESPONSE
                     String responseString;
                     int responseCode = conn.getResponseCode();
-                    LOG.d(LOG_TAG, "response code: " + responseCode);
-                    LOG.d(LOG_TAG, "response headers: " + conn.getHeaderFields());
+                    Log.d(LOG_TAG, "response code: " + responseCode);
+                    Log.d(LOG_TAG, "response headers: " + conn.getHeaderFields());
                     TrackingInputStream inStream = null;
                     try {
                         inStream = getInputStream(conn);
@@ -537,8 +537,8 @@ public class FileTransfer extends CordovaPlugin {
                         safeClose(inStream);
                     }
 
-                    LOG.d(LOG_TAG, "got response from server");
-                    LOG.d(LOG_TAG, responseString.substring(0, Math.min(256, responseString.length())));
+                    Log.d(LOG_TAG, "got response from server");
+                    Log.d(LOG_TAG, responseString.substring(0, Math.min(256, responseString.length())));
 
                     // send request and retrieve response
                     result.setResponseCode(responseCode);
@@ -547,20 +547,20 @@ public class FileTransfer extends CordovaPlugin {
                     context.sendPluginResult(new PluginResult(PluginResult.Status.OK, result.toJSONObject()));
                 } catch (FileNotFoundException e) {
                     JSONObject error = createFileTransferError(FILE_NOT_FOUND_ERR, source, target, conn, e);
-                    LOG.e(LOG_TAG, error.toString(), e);
+                    Log.e(LOG_TAG, error.toString(), e);
                     context.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
                 } catch (IOException e) {
                     JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, conn, e);
-                    LOG.e(LOG_TAG, error.toString(), e);
-                    LOG.e(LOG_TAG, "Failed after uploading " + totalBytes + " of " + fixedLength + " bytes.");
+                    Log.e(LOG_TAG, error.toString(), e);
+                    Log.e(LOG_TAG, "Failed after uploading " + totalBytes + " of " + fixedLength + " bytes.");
                     context.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
                 } catch (JSONException e) {
-                    LOG.e(LOG_TAG, e.getMessage(), e);
+                    Log.e(LOG_TAG, e.getMessage(), e);
                     context.sendPluginResult(new PluginResult(PluginResult.Status.JSON_EXCEPTION));
                 } catch (Throwable t) {
                     // Shouldn't happen, but will
                     JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, conn, t);
-                    LOG.e(LOG_TAG, error.toString(), t);
+                    Log.e(LOG_TAG, error.toString(), t);
                     context.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
                 } finally {
                     synchronized (activeRequests) {
@@ -568,6 +568,7 @@ public class FileTransfer extends CordovaPlugin {
                     }
 
                     if (conn != null) {
+                        // Revert back to the proper verifier and socket factories
                         // Revert back to the proper verifier and socket factories
                         if (trustEveryone && useHttps) {
                             HttpsURLConnection https = (HttpsURLConnection) conn;
@@ -636,7 +637,7 @@ public class FileTransfer extends CordovaPlugin {
             SSLSocketFactory newFactory = sc.getSocketFactory();
             connection.setSSLSocketFactory(newFactory);
         } catch (Exception e) {
-            LOG.e(LOG_TAG, e.getMessage(), e);
+            Log.e(LOG_TAG, e.getMessage(), e);
         }
         return oldFactory;
     }
@@ -671,7 +672,7 @@ public class FileTransfer extends CordovaPlugin {
                 }
             // IOException can leave connection object in a bad state, so catch all exceptions.
             } catch (Throwable e) {
-                LOG.w(LOG_TAG, "Error getting HTTP status code from connection.", e);
+                Log.w(LOG_TAG, "Error getting HTTP status code from connection.", e);
             }
         }
 
@@ -693,7 +694,7 @@ public class FileTransfer extends CordovaPlugin {
             if(body != null)
             {
                 error.put("body", body);
-            }
+            }   
             if (httpStatus != null) {
                 error.put("http_status", httpStatus);
             }
@@ -705,7 +706,7 @@ public class FileTransfer extends CordovaPlugin {
                 error.put("exception", msg);
             }
         } catch (JSONException e) {
-            LOG.e(LOG_TAG, e.getMessage(), e);
+            Log.e(LOG_TAG, e.getMessage(), e);
         }
         return error;
     }
@@ -735,7 +736,7 @@ public class FileTransfer extends CordovaPlugin {
      * @param target            Full path of the file on the file system
      */
     private void download(final String source, final String target, JSONArray args, CallbackContext callbackContext) throws JSONException {
-        LOG.d(LOG_TAG, "download " + source + " to " +  target);
+        Log.d(LOG_TAG, "download " + source + " to " +  target);
 
         final CordovaResourceApi resourceApi = webView.getResourceApi();
 
@@ -744,12 +745,17 @@ public class FileTransfer extends CordovaPlugin {
         final JSONObject headers = args.optJSONObject(4);
 
         final Uri sourceUri = resourceApi.remapUri(Uri.parse(source));
+        // Accept a path or a URI for the source.
+        Uri tmpTarget = Uri.parse(target);
+        final Uri targetUri = resourceApi.remapUri(
+            tmpTarget.getScheme() != null ? tmpTarget : Uri.fromFile(new File(target)));
+
         int uriType = CordovaResourceApi.getUriType(sourceUri);
         final boolean useHttps = uriType == CordovaResourceApi.URI_TYPE_HTTPS;
         final boolean isLocalTransfer = !useHttps && uriType != CordovaResourceApi.URI_TYPE_HTTP;
         if (uriType == CordovaResourceApi.URI_TYPE_UNKNOWN) {
             JSONObject error = createFileTransferError(INVALID_URL_ERR, source, target, null, 0, null);
-            LOG.e(LOG_TAG, "Unsupported URI: " + sourceUri);
+            Log.e(LOG_TAG, "Unsupported URI: " + sourceUri);
             callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
             return;
         }
@@ -786,7 +792,7 @@ public class FileTransfer extends CordovaPlugin {
         }
 
         if (!Boolean.TRUE.equals(shouldAllowRequest)) {
-            LOG.w(LOG_TAG, "Source URL is not in white list: '" + source + "'");
+            Log.w(LOG_TAG, "Source URL is not in white list: '" + source + "'");
             JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, null, 401, null);
             callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.IO_EXCEPTION, error));
             return;
@@ -803,11 +809,6 @@ public class FileTransfer extends CordovaPlugin {
                 if (context.aborted) {
                     return;
                 }
-
-                // Accept a path or a URI for the source.
-                Uri tmpTarget = Uri.parse(target);
-                Uri targetUri = resourceApi.remapUri(
-                        tmpTarget.getScheme() != null ? tmpTarget : Uri.fromFile(new File(target)));
                 HttpURLConnection connection = null;
                 HostnameVerifier oldHostnameVerifier = null;
                 SSLSocketFactory oldSocketFactory = null;
@@ -823,7 +824,7 @@ public class FileTransfer extends CordovaPlugin {
                     file = resourceApi.mapUriToFile(targetUri);
                     context.targetFile = file;
 
-                    LOG.d(LOG_TAG, "Download file:" + sourceUri);
+                    Log.d(LOG_TAG, "Download file:" + sourceUri);
 
                     FileProgressResult progress = new FileProgressResult();
 
@@ -870,7 +871,7 @@ public class FileTransfer extends CordovaPlugin {
                         if (connection.getResponseCode() == HttpURLConnection.HTTP_NOT_MODIFIED) {
                             cached = true;
                             connection.disconnect();
-                            LOG.d(LOG_TAG, "Resource not modified: " + source);
+                            Log.d(LOG_TAG, "Resource not modified: " + source);
                             JSONObject error = createFileTransferError(NOT_MODIFIED_ERR, source, target, connection, null);
                             result = new PluginResult(PluginResult.Status.ERROR, error);
                         } else {
@@ -915,7 +916,7 @@ public class FileTransfer extends CordovaPlugin {
                             safeClose(outputStream);
                         }
 
-                        LOG.d(LOG_TAG, "Saved file: " + target);
+                        Log.d(LOG_TAG, "Saved file: " + target);
 
 
                         // create FileEntry object
@@ -945,28 +946,28 @@ public class FileTransfer extends CordovaPlugin {
                                 result = new PluginResult(PluginResult.Status.OK, fileEntry);
                             } else {
                                 JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, connection, null);
-                                LOG.e(LOG_TAG, "File plugin cannot represent download path");
+                                Log.e(LOG_TAG, "File plugin cannot represent download path");
                                 result = new PluginResult(PluginResult.Status.IO_EXCEPTION, error);
                             }
                         } else {
-                            LOG.e(LOG_TAG, "File plugin not found; cannot save downloaded file");
+                            Log.e(LOG_TAG, "File plugin not found; cannot save downloaded file");
                             result = new PluginResult(PluginResult.Status.ERROR, "File plugin not found; cannot save downloaded file");
                         }
                     }
                 } catch (FileNotFoundException e) {
                     JSONObject error = createFileTransferError(FILE_NOT_FOUND_ERR, source, target, connection, e);
-                    LOG.e(LOG_TAG, error.toString(), e);
+                    Log.e(LOG_TAG, error.toString(), e);
                     result = new PluginResult(PluginResult.Status.IO_EXCEPTION, error);
                 } catch (IOException e) {
                     JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, connection, e);
-                    LOG.e(LOG_TAG, error.toString(), e);
+                    Log.e(LOG_TAG, error.toString(), e);
                     result = new PluginResult(PluginResult.Status.IO_EXCEPTION, error);
                 } catch (JSONException e) {
-                    LOG.e(LOG_TAG, e.getMessage(), e);
+                    Log.e(LOG_TAG, e.getMessage(), e);
                     result = new PluginResult(PluginResult.Status.JSON_EXCEPTION);
                 } catch (Throwable e) {
                     JSONObject error = createFileTransferError(CONNECTION_ERR, source, target, connection, e);
-                    LOG.e(LOG_TAG, error.toString(), e);
+                    Log.e(LOG_TAG, error.toString(), e);
                     result = new PluginResult(PluginResult.Status.IO_EXCEPTION, error);
                 } finally {
                     synchronized (activeRequests) {
@@ -1020,7 +1021,7 @@ public class FileTransfer extends CordovaPlugin {
                             try {
                                 context.connection.disconnect();
                             } catch (Exception e) {
-                                LOG.e(LOG_TAG, "CB-8431 Catch workaround for fatal exception", e);
+                                Log.e(LOG_TAG, "CB-8431 Catch workaround for fatal exception", e);
                             }
                         }
                     }